chore: remove requirements.txt
Concerning the Python package, the dependencies are specified from
the `setup.py` file, while the development dependencies are specified in
the `requirements-dev.txt`. The `requirements.txt` file is only a
duplicate of the requirements specified in `setup.py`, but it doesn't
serve any purpose in our case, so let's get rid of it.